General cleanups.
