+++ osu.Framework/Graphics/Primitives/IPolygon.cs
@@ -7,49 +7,28 @@ namespace osu.Framework.Graphics.Primitives
 {
     public interface IPolygon
     {
+        /// <summary>
+        /// The number of vertices in this polygon.
+        /// </summary>
         int VertexCount { get; }
+
+        /// <summary>
+        /// The number of axes formed by the vertices of this polygon.
+        /// </summary>
         int AxisCount { get; }
 
+        /// <summary>
+        /// Retrieves one of this polygon's vertices.
+        /// </summary>
+        /// <param name="index">The index of the vertex to retrieve.</param>
+        /// <returns>The vertex.</returns>
         Vector2 GetVertex(int index);
+
+        /// <summary>
+        /// Retrieves one of this polygon's axes.
+        /// </summary>
+        /// <param name="index">The index of the axis to retrieve.</param>
+        /// <returns>The axis.</returns>
         Axis GetAxis(int index);
-    }
-
-    public struct Axis
-    {
-        public readonly Vector2 Edge;
-        public readonly Vector2 Normal;
-
-        public Axis(Vector2 firstPoint, Vector2 secondPoint)
-        {
-            Edge = secondPoint - firstPoint;
-            Normal = new Vector2(-Edge.Y, Edge.X);
-        }
-
-        public static bool operator ==(Axis left, Axis right)
-        {
-            return left.Edge == right.Edge;
-        }
-
-        public static bool operator !=(Axis left, Axis right)
-        {
-            return left.Edge != right.Edge;
-        }
-
-        public bool Equals(Axis other)
-        {
-            return Edge.Equals(other.Edge);
-        }
-
-        public override bool Equals(object obj)
-        {
-            if (ReferenceEquals(null, obj))
-                return false;
-            return obj is Axis && Equals((Axis)obj);
-        }
-
-        public override int GetHashCode()
-        {
-            return Edge.GetHashCode();
-        }
     }
 }
